@@ -8,8 +8,8 @@
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
     "axios": "^0.26.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-paginate": "^8.1.3",
     "react-router-dom": "^6.4.2",
     "typescript": "^4.8.4",
